loadClients, get plugins, update voterWeight record for multiple plugins
