+++ plugin-library/index.ts
@@ -1,1 +1,3 @@
 export { updateVoterWeightRecord } from './updateVoterWeightRecord'
+export { loadClient } from './loadClient'
+export { getPlugins } from './getPlugins'
